Extract NameService layer and deduplicate AddressChainViewModel
- Add ResolveNameCase interface in gemcore/cases/name/
- Add NameRepository implementation in data:repositories/name/
- Add NameModule Hilt DI in data:repositories/di/
- Remove duplicate AddressChainViewModel from add_asset:viewmodels
- Update single AddressChainViewModel in recipient:viewmodels to use
  ResolveNameCase instead of direct API client dependency
- Update add_asset:presents and import_wallet to use the shared ViewModel
