+++ android/app/build.gradle.kts
@@ -245,6 +245,7 @@ dependencies {
     implementation(project(":features:settings:settings:presents"))
     implementation(project(":features:settings:settings:viewmodels"))
     implementation(project(":features:recipient:presents"))
+    implementation(project(":features:recipient:viewmodels"))
     implementation(project(":features:nft:presents"))
     implementation(project(":features:wallet-details:presents"))
     implementation(project(":features:bridge:presents"))
