+++ android/features/add_asset/presents/build.gradle.kts
@@ -55,6 +55,7 @@ android {
 dependencies {
     implementation(project(":ui"))
     implementation(project(":features:add_asset:viewmodels"))
+    implementation(project(":features:recipient:viewmodels"))
 
     implementation(libs.hilt.android)
     implementation(libs.hilt.navigation.compose)
