+++ android/gemcore/src/main/kotlin/com/gemwallet/android/cases/name/ResolveNameCase.kt
@@ -1,0 +1,8 @@
+package com.gemwallet.android.cases.name
+
+import com.wallet.core.primitives.Chain
+import com.wallet.core.primitives.NameRecord
+
+interface ResolveNameCase {
+    suspend fun resolveName(name: String, chain: Chain): NameRecord?
+}
